// frontend/src/admin/users/__generated__/AdminAreaUsersListFragment_users.graphql.ts
export type AdminAreaUsersListFragment_users$data = {
  readonly users: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"AdminAreaUserListItemFragment_user">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "AdminAreaUsersListFragment_users";
};
export type AdminAreaUsersListFragment_users$key = {
  readonly " $data"?: AdminAreaUsersListFragment_users$data;
  readonly " $fragmentSpreads": FragmentRefs<"AdminAreaUsersListFragment_users">;
};